salutebot-docs-mcp/requirements.txt
faiss-cpu
llama-index
llama-index-embeddings-huggingface
sentence-transformers
fastmcp
